requirements.txt
ansible-core>=2.20.0,<3.0.0
ansible-pylibssh>=1.3.0,<2.0.0
